+++ approvaltests/src/test/java/org/approvaltests/namer/JUnit4StackTraceNamerTest.java
@@ -1,4 +1,4 @@
-package org.approvaltests.namer.tests;
+package org.approvaltests.namer;
 
 import org.junit.Test;
