+++ approvaltests/src/test/java/org/approvaltests/namer/TestNgStackTraceNamerTest.java
@@ -1,4 +1,4 @@
-package org.approvaltests.namer.tests;
+package org.approvaltests.namer;
 
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
